Application database improvements, dev environment improvements - Fixed newlines in DB secrets - Added database service - Modified a few TLS certs moving between machines - Shifted utility scripts to a dispatch model with orchestrate.sh - Fixed namespace divisions between dev-tools and app - fixed internal domain for DNS - added minikube as a dev setup, still have functionality for k3d for   now - removed database artifact from application
Next step will be to integrate and test with mana connectivity to
database

+++ bin/security/cryptid.sh
@@ -12,6 +12,9 @@ for file in $targets; do
         rm $tmp
         continue
     fi
+
     mv "$tmp" "$target"
-    rm "$file"
+    if [[ -s "$target" ]]; then
+        rm "$file"
+    fi
 done
